Fix docs kotlin configuration metadata location
